runtime/rust: move tests modules to their own file
